Move permission specs into spec/integration/permissions
